AWS/Raspberrypi/Script/requirements.txt
boto3
python-dotenv
opencv-python
awsiotsdk
